Allow Unsized types in BorrowMutex
This means we have to store fat pointers inside the BorrowMutex.
And we can't it do with AtomicPtr anymore - the atomics for 128-bits
or more are not guaranteed on every arch.

We provide an internal AtomicFatPtr which provides Atomic's API but
is essentially a spinlock at the moment. It can be later extended
to be a real atomic on supported architectures and supported
size_of::<*mut T>().