Add optional no_std feature
"std" is now a feature; enabled by default.
This means we drop the dependency on thiserror - Display trait is
implemented manually and the std Error trait is implemented only
with the std feature.

Most importantly, no_std doesn't have an abort() in stable rust,
and instead we just panic. This can be safe if the panic strategy
is set to `abort`, which is now asserted at compile-time.