Acknowledge the BorrowMutex is unsound
Unfortunately.

The mem::forget() or similar can be called on a LendGuard to
obtain multiple mutable references to the same object.
This is now described in README.md.

Also, to slightly contain the scope of UB, the LendGuard can be
now dropped without ever being polled, and any future lends will
work fine.