atomic_waker: Fix a race condition if the waking thread goes to sleep
The polling side was correctly seeing the WAKING state (but not
AWOKEN yet) and was not touching the waker memory that was still
being accessed, but it was resetting the state, and wouldn't be
so safe on the subsequent call. This wasn't taken into account so
far, fix it now.